Update models, app and .gitignore
Remove models from the app.py and put them in modesl.py. Updated
.gitignore to not track .vscode/
