+++ app.py
@@ -1,14 +1,15 @@
 import os
 from datetime import datetime, date
-from flask import Flask, render_template, url_for, redirect, request, flash, send_from_directory
+from flask import Flask, render_template, url_for, redirect, \
+                  request, flash, send_from_directory
 from flask_script import Manager, Shell
 from flask_bootstrap import Bootstrap
 from flask_moment import Moment
-
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate, MigrateCommand
 from werkzeug.security import generate_password_hash, check_password_hash
-from flask_login import LoginManager, UserMixin, login_user, login_required, logout_user
+from flask_login import LoginManager, UserMixin, login_user,\
+                        login_required, logout_user
 from nameform import NameForm, LoginForm
 '''Imported modules'''
 basedir = os.path.abspath(os.path.dirname(__file__))
@@ -17,7 +18,8 @@ basedir = os.path.abspath(os.path.dirname(__file__))
 app = Flask(__name__)
 
 # Database Config
-app.config['SQLALCHEMY_DATABASE_URI']='sqlite:///' + os.path.join(basedir,'sqlite_product_location')
+app.config['SQLALCHEMY_DATABASE_URI'] = \
+            'sqlite:///' + os.path.join(basedir, 'sqlite_product_location')
 app.config['SQLALCHEMY_COMMIT_ON_TEARDOWN'] = True
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SECRET_KEY'] = 'some secret'
@@ -55,43 +57,6 @@ def logout():
     return redirect(url_for('login'))
 
 
-# Models
-class StoreRetrieve(db.Model):
-    __tablename__ = 'Store_Retrieve'
-    location = db.Column(db.String(80), primary_key=True, unique=True)
-    sku = db.Column(db.String(80))
-    date = db.Column(db.String(80))
-
-#    def __init__(self, location, sku, date):
-#        self.location = location
-#        self.sku = sku
-#        self.date = date
-
-    #def __repr__(self):
-        #return '<StoreRetrieve %r>' % self.location
-
-class User(UserMixin, db.Model):
-    __tablename__ = 'users'
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True)
-    password_hash = db.Column(db.String(128))
-
-    @property
-    def password(self):
-        raise AttributeError('password is not a readable attribute')
-
-    @password.setter
-    def password(self, password):
-        self.password_hash = generate_password_hash(password)
-
-    def verify_password(self, password):
-        return check_password_hash(self.password_hash, password)
-
-    #def __repr__(self):
-        #return '<User %r>' % self.username
-
-
-# Imports 4 objects for the database.
 def make_shell_context():
     return dict(app=app, db=db, StoreRetrieve=StoreRetrieve, User=User)
 
@@ -110,9 +75,11 @@ def internal_server_error(e):
     return render_template('500.html', current_time=datetime.utcnow()), 500
 '''Two error pages '''
 
+
 @app.route('/')
 def index():
     return render_template('index.html', current_time=datetime.utcnow())
+
 
 # Page that contains list of warehouse locations, skus and dates
 @app.route('/list')
@@ -121,9 +88,15 @@ def list():
     order = StoreRetrieve.query.order_by(StoreRetrieve.date).all()
     return render_template('list.html', current_time=datetime.utcnow(),
                            order=order)
+
+
 @app.route('/loaderio-50773513826b6129054099d8691c796c/')
 def loaderio():
-    return send_from_directory(os.path.join(os.path.dirname('loaderio-50773513826b6129054099d8691c796c')),'loaderio-50773513826b6129054099d8691c796c.txt')
+    return send_from_directory(
+        os.path.join(os.path.dirname(
+                'loaderio-50773513826b6129054099d8691c796c')),
+                'loaderio-50773513826b6129054099d8691c796c.txt')
+
 
 # Second route retrieve page
 @app.route('/retrieve', methods=['GET', 'POST'])
@@ -153,22 +126,23 @@ filtered by selected proudct, and ordered by date. "older_date" object variable 
 @app.route('/store', methods=['GET', 'POST'])
 @login_required
 def store():
-    #unracked_location = ['A8-2L', 'A8-2R', 'A9-2L', 'A9-2R', 'A13-2L', 'A13-2R']
+    # unracked_location = ['A8-2L', 'A8-2R', 'A9-2L', 'A9-2R', 'A13-2L', 'A13-2R']
 
-    form = NameForm() # instance of NameForm() object
-    empty_bay = StoreRetrieve.query.filter(StoreRetrieve.sku.startswith('Empty')).all() # empty_bay is a 'list' type containing all the 'Empty' skus available
+    form = NameForm()  # instance of NameForm() object
+    empty_bay = StoreRetrieve.query.filter(StoreRetrieve.sku.startswith('Empty')).all()  # empty_bay is a 'list' type containing all the 'Empty' skus available
 
     top_location = [top.location for top in empty_bay if top.location.endswith('3L') or top.location.endswith('3R')]
     # print("print length:", len(top_location), top_location)
 
     third_level_s = [special.location for special in empty_bay if special.location.endswith('3LS') or special.location.endswith('3RS')]
 
-    remaining_location = [remaining.location for remaining in empty_bay if not remaining.location.endswith('3L')
-                          and not remaining.location.endswith('3R') and not remaining.location.endswith('3LS') and
-                          not remaining.location.endswith('3RS')]
-    unracked_location = ['A8-2L', 'A8-2R', 'A9-2L', 'A9-2R', 'A13-2L', 'A13-2R', 'A8-3LS', 'A9-3RS', 'A13-3LS', 'A13-3RS']
-
-
+    remaining_location = [
+        remaining.location for remaining in empty_bay if not remaining.location.endswith('3L') and not remaining.location.endswith('3R') and not remaining.location.endswith('3LS') and
+                          not remaining.location.endswith('3RS')
+                          ]
+    unracked_location = [
+        'A8-2L', 'A8-2R', 'A9-2L', 'A9-2R', 'A13-2L', 'A13-2R', 'A8-3LS', 'A9-3RS', 'A13-3LS', 'A13-3RS'
+        ]
 
     try:
         if request.method == 'POST' and (form.select.data == '464 10lbs' or form.select.data == '444 10lbs') and top_location:
@@ -194,7 +168,6 @@ def store():
             print('elif')
             return redirect(url_for('store'))
 
-
     except:
         flash('No more space for {}'.format(form.select.data))
 
@@ -212,11 +185,8 @@ def store():
                     flash('second try')
                     return redirect(url_for('store'))
 
-
-
     except:
         flash('locations are full')
-
 
     try:
         if request.method == 'POST' and form.select.data != '464 10lbs' and (top_location or not top_location):
@@ -251,6 +221,6 @@ def login():
 
 if __name__ == '__main__':
     app.debug = True
-    app.run('0.0.0.0', port=5000) # to run it locally
-    #port = int(os.environ.get("PORT", 5000)) # to run it on production
+    app.run('0.0.0.0', port=5000)  # to run it locally
+    # port = int(os.environ.get("PORT", 5000)) # to run it on production
     app.run(host='0.0.0.0', port=port)
