+++ models.py
@@ -1,3 +1,5 @@
+import os
+
 basedir = os.path.abspath(os.path.dirname(__file__))
 app = Flask(__name__)
 db = SQLAlchemy(app)
@@ -20,4 +22,43 @@ class StoreRetrieve(db.Model):
         self.date = date
 
     def __repr__(self):
-        return('The location is %s' % self.location)
+        return 'The location is %s' % self.location
+
+
+class StoreRetrieve(db.Model):
+    __tablename__ = 'Store_Retrieve'
+    location = db.Column(db.String(80), primary_key=True, unique=True)
+    sku = db.Column(db.String(80))
+    date = db.Column(db.String(80))
+
+#    def __init__(self, location, sku, date):
+#        self.location = location
+#        self.sku = sku
+#        self.date = date
+
+    #def __repr__(self):
+        #return '<StoreRetrieve %r>' % self.location
+
+
+class User(UserMixin, db.Model):
+    __tablename__ = 'users'
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(80), unique=True)
+    password_hash = db.Column(db.String(128))
+
+    @property
+    def password(self):
+        raise AttributeError('password is not a readable attribute')
+
+    @password.setter
+    def password(self, password):
+        self.password_hash = generate_password_hash(password)
+
+    def verify_password(self, password):
+        return check_password_hash(self.password_hash, password)
+
+    #def __repr__(self):
+        #return '<User %r>' % self.username
+
+
+# Imports 4 objects for the database.
